ada: Tweak CPU affinity handling
The primary motivation for this change is making the taskset command
line tool work as expected for tasking programs that don't use features
from section D.16 of the Ada reference manual. A couple of components
are added to the ATCB record to make it possible to tell values that
come from explicit aspects and subprogram calls from values that are
inherited from activating tasks.

gcc/ada/ChangeLog:

	* libgnarl/s-mudido__affinity.adb (Unchecked_Set_Affinity): Set new
	ATCB component.
	* libgnarl/s-taprop__linux.adb (Create_Task): Only set CPU affinity
	when required.
	(Requires_Affinity_Change): New subprogram.
	(Set_Task_Affinity): Likewise.
	* libgnarl/s-tarest.adb (Create_Restricted_Task): Adapt to
	Initialize_ATCB change.
	* libgnarl/s-taskin.adb (Initialize_ATCB): Update parameter list.
	Record whether aspects were explicitly specified.
	* libgnarl/s-taskin.ads (Common_ATCB): Add component.
	* libgnarl/s-tassta.adb (Create_Task): Update call to Initialize_ATCB.
	* libgnarl/s-tporft.adb (Register_Foreign_Thread): Likewise.